move logs test helper to logs test
