+++ otlp/common_testhelpers.go
@@ -5,18 +5,12 @@ import (
 	"compress/gzip"
 	"errors"
 	"strings"
-	"time"
 
 	"github.com/golang/protobuf/proto"
-	common "github.com/honeycombio/husky/proto/otlp/common/v1"
-	resource "github.com/honeycombio/husky/proto/otlp/resource/v1"
 	"github.com/klauspost/compress/zstd"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/reflect/protoreflect"
 	"google.golang.org/protobuf/runtime/protoiface"
-
-	collectorlogs "github.com/honeycombio/husky/proto/otlp/collector/logs/v1"
-	logs "github.com/honeycombio/husky/proto/otlp/logs/v1"
 )
 
 // The collector<signal>.Export<signal>ServiceRequest structs generated from proto
@@ -89,56 +83,4 @@ func testCaseNameForEncoding(encoding string) string {
 	} else {
 		return encoding
 	}
-}
-
-// Build an OTel Logs request. Provide a valid OTel traceID and spanID, a time for the log entry, and a service name.
-func buildExportLogsServiceRequest(traceID []byte, spanID []byte, startTimestamp time.Time, testServiceName string) *collectorlogs.ExportLogsServiceRequest {
-	req := &collectorlogs.ExportLogsServiceRequest{
-		ResourceLogs: []*logs.ResourceLogs{{
-			Resource: &resource.Resource{
-				Attributes: []*common.KeyValue{{
-					Key: "resource_attr",
-					Value: &common.AnyValue{
-						Value: &common.AnyValue_StringValue{StringValue: "resource_attr_val"},
-					},
-				}, {
-					Key: "service.name",
-					Value: &common.AnyValue{
-						Value: &common.AnyValue_StringValue{StringValue: testServiceName},
-					},
-				}},
-			},
-			ScopeLogs: []*logs.ScopeLogs{{
-				Scope: &common.InstrumentationScope{
-					Name:    "instr_scope_name",
-					Version: "instr_scope_version",
-					Attributes: []*common.KeyValue{
-						{
-							Key: "scope_attr",
-							Value: &common.AnyValue{
-								Value: &common.AnyValue_StringValue{StringValue: "scope_attr_val"},
-							},
-						},
-					},
-				},
-				LogRecords: []*logs.LogRecord{{
-					TraceId:        traceID,
-					SpanId:         spanID,
-					TimeUnixNano:   uint64(startTimestamp.Nanosecond()),
-					SeverityText:   "test_severity_text",
-					SeverityNumber: logs.SeverityNumber_SEVERITY_NUMBER_DEBUG,
-					Attributes: []*common.KeyValue{
-						{
-							Key: "span_attr",
-							Value: &common.AnyValue{
-								Value: &common.AnyValue_StringValue{StringValue: "span_attr_val"},
-							},
-						},
-					},
-				}},
-			}},
-		}},
-	}
-
-	return req
 }
